Move definitions to symbols
